Binance_bot/Eth_model/requirements.txt
python-binance
TA-Lib
numpy
websocket-client
